requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
drf-extensions==0.7.1
pytz==2021.1
sqlparse==0.4.1
